+++ src/main/java/com/achobeta/domain/schedule/controller/InterviewScheduleController.java
@@ -1,14 +1,14 @@
-package com.achobeta.domain.interview.controller;
+package com.achobeta.domain.schedule.controller;
 
 import com.achobeta.common.SystemJsonResponse;
-import com.achobeta.domain.interview.model.dto.ScheduleDTO;
-import com.achobeta.domain.interview.model.dto.ScheduleUpdateDTO;
-import com.achobeta.domain.interview.model.vo.ParticipationDetailVO;
-import com.achobeta.domain.interview.model.vo.ScheduleDetailVO;
-import com.achobeta.domain.interview.model.vo.ScheduleResumeVO;
-import com.achobeta.domain.interview.model.vo.UserSituationVO;
-import com.achobeta.domain.interview.service.InterviewScheduleService;
-import com.achobeta.domain.interview.service.InterviewerService;
+import com.achobeta.domain.schedule.model.dto.ScheduleDTO;
+import com.achobeta.domain.schedule.model.dto.ScheduleUpdateDTO;
+import com.achobeta.domain.schedule.model.vo.ParticipationDetailVO;
+import com.achobeta.domain.schedule.model.vo.ScheduleDetailVO;
+import com.achobeta.domain.schedule.model.vo.ScheduleResumeVO;
+import com.achobeta.domain.schedule.model.vo.UserSituationVO;
+import com.achobeta.domain.schedule.service.InterviewScheduleService;
+import com.achobeta.domain.schedule.service.InterviewerService;
 import com.achobeta.domain.recruit.service.ActivityParticipationService;
 import com.achobeta.domain.recruit.service.RecruitmentActivityService;
 import com.achobeta.domain.users.context.BaseContext;
