+++ src/main/java/com/achobeta/domain/schedule/model/dto/ScheduleUpdateDTO.java
@@ -1,4 +1,4 @@
-package com.achobeta.domain.interview.model.dto;
+package com.achobeta.domain.schedule.model.dto;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
@@ -7,14 +7,14 @@ import lombok.Data;
  * Created With Intellij IDEA
  * Description:
  * User: 马拉圈
- * Date: 2024-07-25
- * Time: 23:15
+ * Date: 2024-07-26
+ * Time: 0:21
  */
 @Data
-public class ScheduleDTO {
+public class ScheduleUpdateDTO {
 
-    @NotNull(message = "用户活动参与票据不能为空")
-    private Long participationId;
+    @NotNull(message = "面试预约 id 不能为空")
+    private Long scheduleId;
 
     @NotNull(message = "开始时间不能为空")
     private Long startTime;
