+++ src/main/java/com/achobeta/domain/schedule/model/entity/InterviewSchedule.java
@@ -1,4 +1,4 @@
-package com.achobeta.domain.interview.model.entity;
+package com.achobeta.domain.schedule.model.entity;
 
 import com.achobeta.common.base.BaseIncrIDEntity;
 import com.baomidou.mybatisplus.annotation.TableName;
