+++ src/main/java/com/achobeta/domain/schedule/model/vo/ParticipationScheduleVO.java
@@ -1,6 +1,5 @@
-package com.achobeta.domain.interview.model.vo;
+package com.achobeta.domain.schedule.model.vo;
 
-import com.achobeta.domain.recruit.model.vo.TimePeriodVO;
 import com.achobeta.domain.student.model.vo.SimpleStudentVO;
 import lombok.Data;
