+++ src/main/java/com/achobeta/domain/schedule/model/vo/UserParticipationVO.java
@@ -1,7 +1,6 @@
-package com.achobeta.domain.interview.model.vo;
+package com.achobeta.domain.schedule.model.vo;
 
 import com.achobeta.domain.recruit.model.vo.TimePeriodVO;
-import com.achobeta.domain.student.model.vo.SimpleStudentVO;
 import lombok.Data;
 
 import java.util.List;
