+++ src/main/java/com/achobeta/domain/schedule/service/impl/InterviewerServiceImpl.java
@@ -1,11 +1,11 @@
-package com.achobeta.domain.interview.service.impl;
+package com.achobeta.domain.schedule.service.impl;
 
 import com.achobeta.common.enums.GlobalServiceStatusCode;
 import com.achobeta.exception.GlobalServiceException;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.achobeta.domain.interview.model.entity.Interviewer;
-import com.achobeta.domain.interview.service.InterviewerService;
-import com.achobeta.domain.interview.model.dao.mapper.InterviewerMapper;
+import com.achobeta.domain.schedule.model.entity.Interviewer;
+import com.achobeta.domain.schedule.service.InterviewerService;
+import com.achobeta.domain.schedule.model.dao.mapper.InterviewerMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
