+++ src/main/resources/mapper/interview/InterviewScheduleMapper.xml
@@ -2,9 +2,9 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.achobeta.domain.interview.model.dao.mapper.InterviewScheduleMapper">
+<mapper namespace="com.achobeta.domain.schedule.model.dao.mapper.InterviewScheduleMapper">
 
-    <resultMap id="BaseResultMap" type="com.achobeta.domain.interview.model.entity.InterviewSchedule">
+    <resultMap id="BaseResultMap" type="com.achobeta.domain.schedule.model.entity.InterviewSchedule">
             <id property="id" column="id" jdbcType="BIGINT"/>
             <result property="participationId" column="participation_id" jdbcType="BIGINT"/>
             <result property="startTime" column="start_time" jdbcType="TIMESTAMP"/>
