+++ src/main/resources/mapper/interview/InterviewerMapper.xml
@@ -2,9 +2,9 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.achobeta.domain.interview.model.dao.mapper.InterviewerMapper">
+<mapper namespace="com.achobeta.domain.schedule.model.dao.mapper.InterviewerMapper">
 
-    <resultMap id="BaseResultMap" type="com.achobeta.domain.interview.model.entity.Interviewer">
+    <resultMap id="BaseResultMap" type="com.achobeta.domain.schedule.model.entity.Interviewer">
             <id property="id" column="id" jdbcType="BIGINT"/>
             <result property="managerId" column="manager_id" jdbcType="BIGINT"/>
             <result property="scheduleId" column="schedule_id" jdbcType="BIGINT"/>
